reorganize by moving files to judges
