+++ components/judges/judgingForm.tsx
@@ -1,6 +1,6 @@
-import ScoreInput from './scoreInput';
+import ScoreInput from '.././scoreInput';
 import { Form, Input, Button } from 'antd';
-import { JudgingFormFields } from '../types/client';
+import { JudgingFormFields } from '../../types/client';
 import { useEffect } from 'react';
 const { TextArea } = Input;
